requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.101.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
pydantic==2.2.1
pydantic_core==2.6.1
pypika-tortoise==0.1.6
python-decouple==3.8
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
